+++ Tetris/LineGoal.java
@@ -4,14 +4,14 @@ public class LineGoal extends Goal {
     private GoalState goalState = GoalState.NONE;
     private int goalLines;
     private int lines;
-    private int timeMillis;
+    private long timeMillis;
 
     public LineGoal(int goalLines) {
         this.goalLines = goalLines;
     }
 
     @Override
-    protected GoalState calculate(int timeMillis, int lines) {
+    protected GoalState calculate(long timeMillis, int lines) {
         this.timeMillis = timeMillis;
         this.lines = lines;
